requirements.txt
openai == 1.3.7
python-dotenv == 1.0.0
langchain == 0.0.345
chromadb == 0.4.18
tiktoken == 0.5.2